creating and clearing relation

+++ REST/Rest-Blog/blog-BE/controllers/feed.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator');
 const Post = require('../models/post');
+const User = require("../models/user");
 const fs = require('fs');
 const path = require('path');
 
@@ -50,17 +51,28 @@ exports.createPosts = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
+
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: "Kapil" },
+        creator: req.userId,
     });
     post.save()
+        .then(() => {
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
         .then(result => {
             res.status(201).json({
                 message: "Post created successfully!",
-                post: result
+                post: post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -121,6 +133,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error("Not authorized.");
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -149,14 +166,25 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error("Not authorized.");
+                error.statusCode = 403;
+                throw error;
+            }
             //Check loggedin user
-
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
-        .then(result => {
-            console.log(result);
+        .then(() => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(() => {
             res.status(200).json({ message: "Deleted the post." })
+
         })
         .catch(err => {
             if (!err.statusCode) {
